+++ lizard_measure/management/commands/import_krw_portaal.py
@@ -291,11 +291,11 @@ def import_measure_types(filename):
             extra_kwargs=extra_kwargs,
         )
 
-        # Add the units
-        units = rec['eenheid'].split(', ')
-        for u_str in units:
-            unit_obj = Unit.objects.get_or_create(code=u_str)[0]
-            measure_type.units.add(unit_obj)
+#       # Add the units
+#       units = rec['eenheid'].split(', ')
+#       for u_str in units:
+#           unit_obj = Unit.objects.get_or_create(code=u_str)[0]
+#           measure_type.units.add(unit_obj)
 
 
 def import_waterbodies(wb_import_settings):
@@ -483,7 +483,6 @@ def import_measures(filename):
             'aggregation_type': Measure.AGGREGATION_TYPE_MIN,
             'description': rec['toelichting'],
             'value': rec['matomv'],
-            'unit': unit,
             'investment_costs': rec['investkosten'],
             'exploitation_costs': rec['exploitkosten'],
             'executive': None,
@@ -576,7 +575,7 @@ def import_measures(filename):
             funding_organization.save()
 
 
-def update_measures(filename):
+def update_measures(filename, typefilename):
     # Fast retrieval instead of separate get() calls
     original_measures = dict([(m.ident, m) for m in Measure.objects.all()])
 
@@ -615,7 +614,6 @@ def update_measures(filename):
                 measure_period.valid = True
                 measure_period.save()
 
-        corresponding_measure.unit = unit
         corresponding_measure.value = rec['matomvbrus']
         corresponding_measure.period = measure_period
         corresponding_measure.save()
@@ -666,6 +664,23 @@ def update_measures(filename):
     for m in Measure.objects.filter(categories=n2000_old):
         m.categories.remove(n2000_old)
         m.categories.add(n2000_new)
+    logger.info(
+        'Replaced measure categories by categories with correct names.',
+    )
+
+    # Adding correct units according to use case
+    typefile = open(typefilename)
+    unit_map = dict(line.split() for line in typefile.readlines())
+    typefile.close()
+    for k, v in unit_map.items():
+        measure_type = MeasureType.objects.get(code=k)
+        measure_type.unit = Unit.objects.get(code=v)
+        measure_type.save()
+    logger.info(
+        'Updated %s MeasureTypes with units from use case',
+        len(unit_map),
+    )
+    
 
 
 class Command(BaseCommand):
@@ -761,8 +776,10 @@ class Command(BaseCommand):
     def _update(self, import_path):
 
         # Update measures
-        update_measures(os.path.join(import_path, 'maatregelen_update.xml'))
-
+        update_measures(
+            filename=os.path.join(import_path, 'maatregelen_update.xml'),
+            typefilename=os.path.join(import_path, 'maatregeltype_eenheid.txt'),
+        )
 
 
     @transaction.commit_on_success